% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_mean.R
\name{c_mean}
\alias{c_mean}
\title{c_mean() computes the cluster mean for given variable, and a given clustering variable}
\usage{
c_mean(x, j)
}
\arguments{
\item{x}{a numeric vector of a single variable}
}
\value{
a numeric vector
}
\description{
c_mean() computes the cluster mean for given variable, and a given clustering variable
}
\details{
simple wrapper for `ave(variable, group)` function, to estimate cluster means
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(cluster_mean =  r4sda::c_mean(x, id_j)) \%>\%
select(x, id_j) \%>\%
unique() \%>\%
print
}
