% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_median.R
\name{c_median}
\alias{c_median}
\title{c_median() computes the cluster median for given variable, and a given clustering variable}
\usage{
c_median(x, j)
}
\arguments{
\item{x}{a numeric vector of a single variable}

\item{j}{a numeric vector that indexes values into groups}
}
\value{
a numeric vector
}
\description{
c_median() computes the cluster median for given variable, and a given clustering variable
}
\details{
simple wrapper for `ave(variable, group)` function, to estimate cluster means
}
\examples{
example_data <- data.frame(x = c(1,2,3,4,5,6), j = c(1,1,1,2,2,2))
library(dplyr)
example_data \%>\%
mutate(cluster_median =  c_median(x, j)) \%>\%
select(x, j) \%>\%
unique() \%>\%
print
}
