% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_sd.R
\name{c_sd}
\alias{c_sd}
\title{c_sd() computes the cluster standard deviation for given variable, and a given clustering variable}
\usage{
c_sd(x, j)
}
\arguments{
\item{x}{a numeric vector of a single variable}
}
\value{
a numeric vector
}
\description{
c_sd() computes the cluster standard deviation for given variable, and a given clustering variable
}
\details{
simple wrapper for `ave(variable, group)` function, to estimate cluster standard deviations
}
\examples{
example_data <- data.frame(x = c(1,2,3,4,5,6), j = c(1,1,1,2,2,2))
library(dplyr)
example_data \%>\%
mutate(cluster_sd =  c_sd(x, j)) \%>\%
select(x, j) \%>\%
unique() \%>\%
print
}
