% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wmean.R
\name{c_wmean}
\alias{c_wmean}
\title{c_wmean() computes the cluster weighted mean for a given variable, and a given clustering variable}
\usage{
c_wmean(x, j)
}
\arguments{
\item{x}{a numeric vector of a single variable}

\item{wi}{a numeric vector for within cluster weights}
}
\value{
a numeric vector
}
\description{
c_wmean() computes the cluster weighted mean for a given variable, and a given clustering variable
}
\details{
simple wrapper for `ave(variable, group)` function couple with matrixStats::weightedMean(), to estimate cluster means weighted by within cluster weights
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(cluster_mean =  c_wmean(x, id_j)) \%>\%
select(x, id_j) \%>\%
unique() \%>\%
print
}
