% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caterpillar_plot.R
\name{caterpillar_plot}
\alias{caterpillar_plot}
\title{caterpillar_plot() it plots the rank of random effects across clusters, with ther point estimates and their uncertainty (2*SD), around it points estimates}
\usage{
caterpillar_plot(mlm_model)
}
\arguments{
\item{mlm_model}{it requires an mlm_model, generated by lme4. If this is a null, the plot display the unconditioned random effect. If the model is not a null model, then the plot display the expected random effects given the model.}
}
\value{
a plot, a `ggplot2` object.
}
\description{
caterpillar_plot() it plots the rank of random effects across clusters, with ther point estimates and their uncertainty (2*SD), around it points estimates
}
\details{
Given the return object is a common `ggplot2` user can further customize the figure, overwritting or defining parameter for the plot such as colours, axis lenght, and coord_fixed(ratio=.xx)
}
\examples{

lme4::lmer(y ~ 1 + (1|id_j), data = clustered_data, REML = FALSE) \%>\%
caterpillar_plot(.)



}
