% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_variables.R
\name{center_variables}
\alias{center_variables}
\title{center_variables() computes the centered variables for mixed models}
\usage{
center_variables(data, variable, name)
}
\arguments{
\item{data}{a data frame with variables id_j (cluster variable), and id_k (higher cluster, e.g., country)}

\item{variable}{variable to be center}

\item{name}{acronym for the generated variables}
}
\value{
a data frame, including the center variables
}
\description{
center_variables() computes the centered variables for mixed models
}
\examples{
library(dplyr)
data_model \%>\%
center_variables(data = ., variable = sex , name = sex) \%>\%
dplyr::glimpse()
}
