% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_location.R
\name{char_location}
\alias{char_location}
\title{char_location() gets the location of a character within a string}
\usage{
char_location(string, char)
}
\arguments{
\item{string}{is a string or text line}

\item{char}{is character we want to locate}
}
\value{
returns a number, which is the location before the character is found within a string
}
\description{
char_location() gets the location of a character within a string
}
\details{
is a simple wrapper of the function `stringr::str_locate`. Is used to generate codebook.
}
\examples{
char_location('this is text', 'x')
}
