% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cluster_id.R
\name{check_cluster_id}
\alias{check_cluster_id}
\title{z_score() it standardized a continous variable}
\usage{
check_cluster_id(data, id_j, id_k)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an standardized variable, as a z score
}
\description{
z_score() it standardized a continous variable
}
\examples{
\\donotrun{

data_example \%>\%
check_cluster_id(
  id_j = 'schools',
  id_k = 'country')
}
}
