% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal.R
\name{decimal}
\alias{decimal}
\title{decimal() takes any numeric variable with decimal places, and returns its aproximation of decimal places to a given number.}
\usage{
decimal(x, k)
}
\arguments{
\item{x}{a numeric vector of a single variable}
}
\value{
a string containing a numeric value, with a given number of decimals
}
\description{
decimal() takes any numeric variable with decimal places, and returns its aproximation of decimal places to a given number.
}
\details{
simple wrapper for `format(round(x, k), nsmall=k)` function, to format decimal numbers
}
\examples{

number_with_decimals <- .5838
number_with_two_decimals <- decimal(number_with_decimals,2)
# result is .58
number_with_two_decimals
}
