% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_turnover.R
\name{first_turnover}
\alias{first_turnover}
\title{first_turnover() computes the last time an element is observed in a list}
\usage{
first_turnover(a)
}
\arguments{
\item{a}{numeric vector such as year, or ordered number}
}
\value{
a numeric vector
}
\description{
first_turnover() computes the last time an element is observed in a list
}
\details{
sequence function to identify a year an ordering index
}
\examples{
library(dplyr)
person_sequence <- stacked_data \%>\% 
                   dplyr::select(year, id_j, id_i, id_ij) \%>\%
                   unique() \%>\%
                   group_by(id_ij) \%>\%
                   summarise(start = min(year), final = first_turnover(year)) \%>\% 
                   mutate(time = final - start + 1)

}
