% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icc.R
\name{get_icc}
\alias{get_icc}
\title{get_icc() it reurns the intra class correlation of an mlm model}
\usage{
get_icc(x)
}
\arguments{
\item{mlm_model}{it requires an mlm_model, generated by lme4. If this is a null, is the percentage of variance between clusters.}
}
\value{
a numeric object.
}
\description{
get_icc() it reurns the intra class correlation of an mlm model
}
\details{
if the mlm_model is a null model it returns the ICC. If is a mlm_model with covariates, it returns the conditional variance between clusters.
}
\examples{

lme4::lmer(y ~ 1 + (1|id_j), data = clustered_data, REML = FALSE) \%>\%
get_icc(.)



}
