% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_item_desc.R
\name{get_item_desc}
\alias{get_item_desc}
\title{get_item_desc() produces a table with descriptives where items or variables are rows and columns are differerent descriptive values. its largely based on the deprecated function skimr::skim_to_wide}
\usage{
get_item_desc(x)
}
\arguments{
\item{x}{a data frame, where rows = observations, and columns = variables}
}
\value{
data frame with with descriptive values
}
\description{
get_item_desc() produces a table with descriptives where items or variables are rows and columns are differerent descriptive values. its largely based on the deprecated function skimr::skim_to_wide
}
\examples{


items_data <- dplyr::select(data_frame, item_1:item_9)
get_item_desc(items_data)

}
