% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lrt_scf.R
\name{get_lrt_scf}
\alias{get_lrt_scf}
\title{get_lrt_scf() compares the output of two MPLUS nested models, and calculates the adjusted LRT}
\usage{
get_lrt_scf(model_0, model_1)
}
\arguments{
\item{model_0}{is de constrained model}

\item{model_1}{is the model withe free parameter or parameters}
}
\value{
a string with results one could paste into the manuscipt
}
\description{
get_lrt_scf() compares the output of two MPLUS nested models, and calculates the adjusted LRT
}
\details{
simple wrapper for LRT adjusted test (see https://www.statmodel.com/chidiff.shtml)
}
\examples{

get_lrt_scf(m00, m01) \%>\% knitr::kable()

}
