% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackknifes_iccs.R
\name{jackknifes_iccs}
\alias{jackknifes_iccs}
\title{jackknifes_iccs() computes jackknifes for ICCS 2009 and ICCS 2016}
\usage{
jackknifes_iccs(data, nrep, jkz, jki, wgt)
}
\arguments{
\item{data}{data frame object which contains the rest of the arguments}

\item{nrep}{number of replicates}

\item{jkz}{vector containing the pseudo strata (75 in ICCS studies)}

\item{jki}{vector selecting which school goes into the replicate}

\item{wgt}{total survey weights}
}
\value{
the original data frame with the jacknnifes replicates added.
}
\description{
jackknifes_iccs() computes jackknifes for ICCS 2009 and ICCS 2016
}
\details{
a loop function to create jacknife replicate weights (author: `merivera@uc.cl`)
}
\examples{
library(dplyr)
data_frame_with_jackknifes <- data_frame \%>\%
                              r4sda::jackknifes_iccs(
                              data = .,
                              nrep = 75,
                              jkz = 'JKZONEZ',
                              jki = 'JKREPS',
                              wgt = 'TOTWGTS')

}
