% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_path.R
\name{local_path}
\alias{local_path}
\title{local_path() returns the parent directory of the working directory.It aids the generation of relative directory to store all code in single folder, and then reproduce the calculations in a Mac OS machine or a Windows machine, or else.}
\usage{
local_path(x)
}
\arguments{
\item{x}{a string with directory location (e.g, '/00_data/', '/01_syntax/', '/02_tables/')}
}
\value{
an absolute route or logical directory for the current machine, adding its root directory
}
\description{
local_path() returns the parent directory of the working directory.It aids the generation of relative directory to store all code in single folder, and then reproduce the calculations in a Mac OS machine or a Windows machine, or else.
}
\examples{


parent_directory <- local_path(getwd())
parent_directory #  this is the current working directory
                 #  it assumes the syntax folder, is within the parent directory
}
