% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsa_weights.R
\name{lsa_weights}
\alias{lsa_weights}
\title{lsa_weights() computes normalized and effective sample size weights for pseudo maximum likelihood mixed models, with large scale assessment data}
\usage{
lsa_weights(data, id_i, id_j, id_k, wt, wi, wj)
}
\arguments{
\item{data}{data frame object which contains the rest of the arguments}

\item{id_i}{a numeric vector of a single variable}

\item{id_j}{a numeric vector of a single variable}

\item{id_k}{a numeric vector of a single variable}

\item{wt}{observations final weight}

\item{wi}{observations weigth within the primary sampling unit (i.e. students eight within the school)}

\item{wj}{cluster weights (i.e. school weights)}
}
\value{
the original data frame with the new weights at the end, where `wa1` nad `wa2` are level 1 and level 2 normalized weights; `wb1` nad `wb2` are level 1 and level 2 effectives sample weights.
}
\description{
lsa_weights() computes normalized and effective sample size weights for pseudo maximum likelihood mixed models, with large scale assessment data
}
\details{
a collection of dplyr::mutate sequences to create normalized and effective sample weights (see Rabe-Hesketh & Skrondal, 2006; Snijder & Bosker, 2012)
}
\examples{
library(dplyr)
data_frame_with_weights <- data_frame \%>\%
                           r4sda::lsa_weights(.,
                             id_i = 'id_i',
                             id_j = 'id_j',
                             id_k = 'id_k',
                             wt = 'wt',
                             wi = 'wi',
                             wj = 'wj' )
}
