% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_score.R
\name{mean_score}
\alias{mean_score}
\title{mean_score() computes the mean for a given set of variables in a row wise manner}
\usage{
mean_score(..., na.rm = TRUE)
}
\arguments{
\item{x, y, z}{are set of variables from a data frame (i.e. a selection of numeric columns)}
}
\value{
a numeric vector
}
\description{
mean_score() computes the mean for a given set of variables in a row wise manner
}
\details{
simple wrapper to produce row wise means (default is na.rm=TRUE)
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(score =  mean_score(x, z, y))


}
