% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_score.R
\name{na_score}
\alias{na_score}
\title{na_score() computes a count of missings across a given set of variables in a row wise manner}
\usage{
na_score(..., na.rm = TRUE)
}
\arguments{
\item{x, y, z}{are set of variables from a data frame (i.e. a selection of numeric columns)}
}
\value{
a numeric vector
}
\description{
na_score() computes a count of missings across a given set of variables in a row wise manner
}
\details{
simple wrapper to produce rowwise sums (default is na.rm=TRUE)
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(na_count =  na_score(x, z, y))


}
