% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_score.R
\name{obs_score}
\alias{obs_score}
\title{obs_score() computes a count of valid observes across a given set of variables in a row wise manner}
\usage{
obs_score(..., na.rm = TRUE)
}
\arguments{
\item{x, y, z}{are set of variables from a data frame (i.e. a selection of numeric columns)}
}
\value{
a numeric vector
}
\description{
obs_score() computes a count of valid observes across a given set of variables in a row wise manner
}
\details{
simple wrapper to produce row wise sums (default is na.rm=TRUE)
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(obs_score =  obs_score(x, z, y))


}
