% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_labels.R
\name{remove_labels}
\alias{remove_labels}
\title{remove_labels() from a data frame}
\usage{
remove_labels(x)
}
\arguments{
\item{data}{data frame with labels}
}
\value{
the original data frame without any labels
}
\description{
remove_labels() from a data frame
}
\examples{
library(dplyr)
data_without_labels <- data_frame \%>\%
                       remove_labels()


}
