% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{reverse() it reverse a score in a linear way, using the minimum and maximum values. Is often use onto likert type response variables, when the original responses are not coded with values that follow the direction of higher values are higher level of the attribute.}
\usage{
reverse(var)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a reverse coded numeric vector
}
\description{
reverse() it reverse a score in a linear way, using the minimum and maximum values. Is often use onto likert type response variables, when the original responses are not coded with values that follow the direction of higher values are higher level of the attribute.
}
\examples{

items_data <- items_data \%>\%
              mutate(item = reverse(original))
dplyr::count(items_data, original, item)

}
