% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senate_weights.R
\name{senate_weights}
\alias{senate_weights}
\title{lsa_weights() computes normalized and effective sample size weights for pseudo maximum likelihood mixed models, with large scale assessment data}
\usage{
senate_weights(data, wt, scale, id_k)
}
\arguments{
\item{data}{data frame object which contains the rest of the arguments}

\item{wt}{observations final weight}

\item{scale}{number to which the weights will be normalized to}

\item{id_k}{a numeric vector of a single variable, to distinguish between countries}
}
\value{
the original data frame with the new senate weights at the end, 'ws'
}
\description{
lsa_weights() computes normalized and effective sample size weights for pseudo maximum likelihood mixed models, with large scale assessment data
}
\details{
a collection of dplyr::mutate sequences to create normalized and effective sample weights (see Rabe-Hesketh & Skrondal, 2006; Snijder & Bosker, 2012)
}
\examples{
library(dplyr)
data_frame_with_weights <- data_frame \%>\%
                           r4sda::lsa_weights(.,
                             scale = 1000,
                             wt = 'wt'
                             id_k = 'id_k')

}
