% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent.R
\name{silent}
\alias{silent}
\title{silent() suppresses and hides other functions messages and warnings. Its ideal use if to wrapped functions within dynamic reports.}
\usage{
silent(x)
}
\arguments{
\item{x}{a function wrapping its inputs}
}
\value{
a the product of the wrapped function without messages and warnings
}
\description{
silent() suppresses and hides other functions messages and warnings. Its ideal use if to wrapped functions within dynamic reports.
}
\examples{

table <- items_data \%>\%
         psych::alpha() \%>\%
         silent()

}
