% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_resp.R
\name{stack_resp}
\alias{stack_resp}
\title{stack_resp() produces a frequency table with percentages where items are rows, responses are attributes of items, and percentages are presented for each response.}
\usage{
stack_resp(x)
}
\arguments{
\item{x}{a data frame, where rows = observations, and columns = variables}
}
\value{
data frame with a frequency table
}
\description{
stack_resp() produces a frequency table with percentages where items are rows, responses are attributes of items, and percentages are presented for each response.
}
\examples{


items_data <- dplyr::select(data_frame, item_1:item_9)
stack_resp(items_data)

}
