% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_score.R
\name{sum_score}
\alias{sum_score}
\title{sum_score() computes the mean for a given set of variables in a row wise manner}
\usage{
sum_score(..., na.rm = TRUE)
}
\arguments{
\item{x, y, z}{are set of variables from a data frame (i.e. a selection of numeric columns)}
}
\value{
a numeric vector
}
\description{
sum_score() computes the mean for a given set of variables in a row wise manner
}
\details{
simple wrapper to produce row wise sums (default is na.rm=TRUE)
}
\examples{
library(dplyr)
data_frame \%>\%
mutate(score =  sum_score(x, z, y))


}
