% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_freq.R
\name{svy_freq}
\alias{svy_freq}
\title{svy_freq() estimates proportion for observed values of a categorical variable, from a survey object.}
\usage{
svy_freq(data, variable)
}
\arguments{
\item{data}{a survey object, or survey data frame (i.e. TSL, JKN, BRR or else)}

\item{variable}{selected varible from the survey object}
}
\value{
a tibble with the estimated proportion, and their 95% confidence interval
}
\description{
svy_freq() estimates proportion for observed values of a categorical variable, from a survey object.
}
\examples{


data_svy \%>\%
svy_freq(
  data = .,
  variable = selected_var)

}
