% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_label.R
\name{variable_label}
\alias{variable_label}
\title{variable_label() gets the variable label from a labelled variable, from a data_frame}
\usage{
variable_label(d)
}
\arguments{
\item{d}{is a variable from a data_frame}
}
\value{
returns a string, which contains the variable label of variable, from a labelled data_frame
}
\description{
variable_label() gets the variable label from a labelled variable, from a data_frame
}
\details{
is a wrapper to extract the label attribute from an object
}
\examples{
variable_label(data_frame$variable)

}
