% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables_table.R
\name{variables_table}
\alias{variables_table}
\title{variables_table() produces a simple codebook, based on the output of a dplyr::glimpse()}
\usage{
variables_table(d)
}
\arguments{
\item{d}{is a data_frame}
}
\value{
returns a data_frame in tibble::tibble format.
}
\description{
variables_table() produces a simple codebook, based on the output of a dplyr::glimpse()
}
\details{
it relies on char_location, variable_label, stringr::str_sub, purr::map_dfr, dplyr::glimpse
}
\examples{
variable_label(data_frame$variable)

}
