% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cgm.R
\name{weighted_cgm}
\alias{weighted_cgm}
\title{weighted_cgm() computes the grand mean centering, while including normalized survey weights}
\usage{
weighted_cgm(data, variable, weight, name)
}
\arguments{
\item{data}{a data frame with variables id_j (cluster variable), and id_k (higher cluster, e.g., country)}

\item{variable}{variable to be center}

\item{name}{acronym for the generated variables}
}
\value{
a data frame, including the center variables
}
\description{
weighted_cgm() computes the grand mean centering, while including normalized survey weights
}
\examples{
library(dplyr)
data_model \%>\%
center_variables(data = ., variable = sex , name = sex) \%>\%
dplyr::glimpse()
}
