% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_resp.R
\name{wide_resp}
\alias{wide_resp}
\title{wide_resp() produces a frequency table in percentages where items are rows and columns are response values}
\usage{
wide_resp(x)
}
\arguments{
\item{x}{a data frame, where rows = observations, and columns = variables}
}
\value{
data frame with a frequency table
}
\description{
wide_resp() produces a frequency table in percentages where items are rows and columns are response values
}
\examples{


items_data <- dplyr::select(data_frame, item_1:item_9)
wide_resp(items_data)

}
