% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_var.R
\name{wide_var}
\alias{wide_var}
\title{wide_var() produces a frequency table with percentages where responses are rows, and columnas are variables.}
\usage{
wide_var(x)
}
\arguments{
\item{x}{a data frame, where rows = observations, and columns = variables}
}
\value{
data frame with a frequency table
}
\description{
wide_var() produces a frequency table with percentages where responses are rows, and columnas are variables.
}
\examples{

items_data <- dplyr::select(data_frame, item_1:item_9)
wide_var(items_data)

}
