% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{z_score() it standardized a continuous variable}
\usage{
z_score(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an standardized variable, as a z score
}
\description{
z_score() it standardized a continuous variable
}
\examples{

items_data <- items_data \%>\%
              mutate(zcore = zscore(original))

}
