% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{zscore() it standardized a continous measure}
\usage{
zscore(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
an standardized variable, as a z score
}
\description{
zscore() it standardized a continous measure
}
\examples{
\\donotrun{
items_data <- items_data \%>\%
              mutate(zcore = zscore(original))
skimr::skim(items_data$zcore)
}
}
