% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_wmean.R
\name{c_wmean}
\alias{c_wmean}
\title{c_wmean() computes the cluster weighted mean for a given variable, and a given clustering variable}
\usage{
c_wmean(x, w, j)
}
\arguments{
\item{x}{a numeric vector of a single variable}

\item{w}{a numeric vector for within cluster weights}

\item{j}{a numeric vector for clusters index}
}
\value{
a numeric vector
}
\description{
c_wmean() computes the cluster weighted mean for a given variable, and a given clustering variable
}
\details{
simple wrapper for  ave(x,j), to estimate cluster means weighted by within cluster weights
}
\examples{
library(dplyr)
library(dplyr)
example_data <- data.frame(
x = c(1,2,3,4,5,6),
w = c(.4,.4,.4,.6,.6,.6),
id_j = c(1,1,1,2,2,2)
)
example_data \%>\%
mutate(cluster_mean =  c_wmean(x, w, id_j)) \%>\%
select(x, id_j, cluster_mean) \%>\%
unique() \%>\%
print
}
