% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_path.R
\name{local_path}
\alias{local_path}
\title{local_path() turns a relative directory into a absolute directory. It aids the generation of relative directory to store all code in single folder, and then reproduce the calculations in a Mac OS machine or a Windows machine, or else. Is simple wrapper for `tools::file_path_as_absolute(x)`, to which it adds a final `/`.}
\usage{
local_path(x)
}
\arguments{
\item{x}{a string with directory location}
}
\value{
an absolute route or logical directory for the current machine
}
\description{
local_path() turns a relative directory into a absolute directory. It aids the generation of relative directory to store all code in single folder, and then reproduce the calculations in a Mac OS machine or a Windows machine, or else. Is simple wrapper for `tools::file_path_as_absolute(x)`, to which it adds a final `/`.
}
\examples{
working_directory <- local_path(getwd())
\donotrun{
working_directory <- local_path(getwd())
working_directory # this is the current working directory
}
}
