% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svy_freq.R
\name{svy_freq}
\alias{svy_freq}
\title{svy_freq() produces proportion for observed values of a numeric variable. It generates expected proportions out of a categorical variable, for each alternative.}
\usage{
svy_freq(data, variable)
}
\arguments{
\item{data}{a survey object, or survey data frame (i.e. TSL, JKN, BRR or else)}

\item{variable}{selected varible from the survey object}
}
\value{
a tibble with the estimated proportion, and their 95% confidence interval
}
\description{
svy_freq() produces proportion for observed values of a numeric variable. It generates expected proportions out of a categorical variable, for each alternative.
}
\examples{


data_svy \%>\%
svy_freq(
  data = .,
  variable = selected_var)

}
