% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_cgm.R
\name{weighted_cgm}
\alias{weighted_cgm}
\title{weighted_cgm() computes the grand mean centering, while including normalized survey weights}
\usage{
weighted_cgm(data, variable, weight, name)
}
\arguments{
\item{data}{a data frame with variables ws (normalized survey weight), and id_k (higher cluster, e.g., country)}

\item{variable}{variable to be center}

\item{weight}{normalized survey weight}

\item{name}{acronym for the generated variables}
}
\value{
a data frame, including the center variables
}
\description{
weighted_cgm() computes the grand mean centering, while including normalized survey weights
}
\examples{
library(dplyr)
data_model \%>\%
weighted_cgm(data = ., variable = sex , weight = ws, name = sex) \%>\%
dplyr::glimpse()
}
